Add ScriptAnalyser checks to build script

+++ build-package-management.ps1
@@ -5,6 +5,23 @@ $OutputFolder = 'out\PackageManagement'
 $SourceFolder = 'src\PackageManagement'
 $DocsFolder = 'docs\PackageManagement'  
 
+Import-Module -Name PSScriptAnalyzer
+
+$Results = Invoke-ScriptAnalyzer -Path src\PackageManagement\AzureAutomationTools.PackageManagement.psm1 -Settings 'CodeFormatting'
+
+if($Results.Count -ne 0){
+    $Results 
+    throw 'Failed CodeFormatting rules'
+}
+
+# excluding PSUseDeclaredVarsMoreThanAssignments as seems to false positive on += to arrays
+$Results = Invoke-ScriptAnalyzer -Path src\PackageManagement\AzureAutomationTools.PackageManagement.psm1 -Settings 'PSGallery' -ExcludeRule 'PSUseDeclaredVarsMoreThanAssignments'
+
+if($Results.Count -ne 0){
+    $Results 
+    throw 'Failed PSGallery rules'
+}
+
 if(Test-Path -Path $OutputFolder -PathType Any){
     Remove-Item -Path $OutputFolder -Recurse
 }
@@ -14,3 +31,4 @@ Copy-Item -Path $SourceFolder\* -Destination $OutputFolder -Recurse
 Import-Module (Resolve-Path $OutputFolder\AzureAutomationTools.PackageManagement.psd1)
 New-ExternalHelp -Path $DocsFolder -OutputPath "$OutputFolder\en-US"
 
+Write-Verbose -Message (Resolve-Path $OutputFolder) -Verbose
